created MapSidebar component
